backend/requirements.txt
# Django et base de données
Django==5.0.8
# psycopg2-binary déjà installé manuellement
python-decouple==3.8
python-dotenv==1.0.0

# Django REST Framework
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Traitement vidéo et média
Pillow==10.4.0
moviepy==1.0.3

# Utilitaires
python-dateutil==2.8.2
pytz==2024.1

# Développement (optionnel)
django-debug-toolbar==4.2.0